Switch to GUI-configured credentials.
As much as I want to trust having a secure key in the YAML config, as it
is encrypted and can only be decrypted during the deployment step which
doesn't offer scripting abilities or random PRs, it just feels safer if
it's not there.  It's also easier to manage through the AppVeyor GUI.

It also makes a test fork easier to manage.