Test to ensure the '/Release' flag installs a specific version.
The key to solving the mysterious malfunction was the fact that
Chocolatey has a known, but not well documented, failure installing from
a nupkg directly instead of installing from a directory (or feed) by
just package name.  That's fine as that's probably a better real-world
test, but in testing you'd might first expect that it would be best to
use the direct path to the nupkg you'd just created!  Alas.