Simply match filenames, rather than ensuring they're not a directories.
The original code here used the '-file' specifier to 'Get-ChildItem' but
due to that feature being missing in PowerShell 2, the
'!$_.PSIsContainer' matcher was added to fix it.

The intention this filtering was to ensure that "containers" (Windows-
slang for a "Directory") were not inadvertently matched.  While it's
super, super defensive to ensure that the '.tar' and '.tar.gz' files
we've downloaded are not directories, I think that file pattern match is
more than sufficient.

Overall, I'd rather not introduce PowerShell 2 syntax merely for making
that overly-defensive move and just count on the fact that a .tar file
and a .tar.gz are not directories, since we control the code that made
them in the first place.